% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_estimation.R
\name{additive_matrix_als_factorize_joint_cov}
\alias{additive_matrix_als_factorize_joint_cov}
\title{Construct the additive component matrix}
\usage{
additive_matrix_als_factorize_joint_cov(object)
}
\arguments{
\item{object}{A fitted object returned by als_factorize_joint_cov().}
}
\value{
An N x M matrix containing alpha_i + beta_j.
}
\description{
Construct the additive component matrix
}
