% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{coef_draws_bootstrap_als}
\alias{coef_draws_bootstrap_als}
\title{Extract coefficient draws from bootstrap ALS object}
\usage{
coef_draws_bootstrap_als(object)
}
\arguments{
\item{object}{A bootstrap object returned by bootstrap_als_factorize_joint_cov().}
}
\value{
A matrix of bootstrap coefficient draws.
}
\description{
Extract coefficient draws from bootstrap ALS object
}
