% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{coef_summary_bootstrap_als}
\alias{coef_summary_bootstrap_als}
\title{Extract coefficient summary table from bootstrap ALS object}
\usage{
coef_summary_bootstrap_als(object)
}
\arguments{
\item{object}{A bootstrap object returned by bootstrap_als_factorize_joint_cov().}
}
\value{
A data.frame of coefficient summaries.
}
\description{
Extract coefficient summary table from bootstrap ALS object
}
