% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareAlignments.R
\name{compareAlignments}
\alias{compareAlignments}
\title{Compare Alignments (Deprecated)}
\usage{
compareAlignments(
  ...,
  names,
  smooth = FALSE,
  se = FALSE,
  min_frac = NULL,
  max_frac = NULL
)
}
\arguments{
\item{...}{Alignment objects.}

\item{names}{Character vector of names.}

\item{smooth}{Logical; add smooth line.}

\item{se}{Logical; add standard error bands.}

\item{min_frac}{Numeric; minimum fraction.}

\item{max_frac}{Numeric; maximum fraction.}
}
\value{
None. Raises an error.
}
\description{
This function is deprecated and will be reimplemented in a
future version.
}
