% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cma.R
\name{consensusEdges}
\alias{consensusEdges}
\alias{consensusEdges,ConsortiumMetabolismAlignment-method}
\title{Get Consensus Edges}
\usage{
consensusEdges(object)

\S4method{consensusEdges}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{A \link{ConsortiumMetabolismAlignment} object
of type \code{"multiple"}.}
}
\value{
A data.frame of consensus edges with prevalence.
}
\description{
Returns the consensus network edges from a multiple
alignment.
}
\section{Methods (by class)}{
\itemize{
\item \code{consensusEdges(ConsortiumMetabolismAlignment)}: Consensus edges from a
multiple \link{ConsortiumMetabolismAlignment}

}}
