% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cma.R
\name{consensusPathways}
\alias{consensusPathways}
\alias{consensusPathways,ConsortiumMetabolismAlignment-method}
\title{Get Consensus Pathways}
\usage{
consensusPathways(object)

\S4method{consensusPathways}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{A \link{ConsortiumMetabolismAlignment} object
of type \code{"multiple"}.}
}
\value{
A data.frame of consensus pathways with
prevalence.
}
\description{
Returns the consensus network pathways from a multiple
alignment.
}
\section{Methods (by class)}{
\itemize{
\item \code{consensusPathways(ConsortiumMetabolismAlignment)}: Consensus pathways from
a multiple \link{ConsortiumMetabolismAlignment}

}}
\examples{
\donttest{
cm1 <- synCM("comm_1", n_species = 3, max_met = 5)
cm2 <- synCM("comm_2", n_species = 4, max_met = 6)
cms <- ConsortiumMetabolismSet(cm1, cm2, name = "test")
cma <- align(cms)
consensusPathways(cma)
}

}
