% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cm.R,
%   R/methods-accessors-cma.R, R/methods-accessors-cms.R
\name{consortia}
\alias{consortia}
\alias{consortia,ConsortiumMetabolism-method}
\alias{consortia,ConsortiumMetabolismAlignment-method}
\alias{consortia,ConsortiumMetabolismSet-method}
\title{Get the Consortia}
\usage{
consortia(object)

\S4method{consortia}{ConsortiumMetabolism}(object)

\S4method{consortia}{ConsortiumMetabolismAlignment}(object)

\S4method{consortia}{ConsortiumMetabolismSet}(object)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismSet} object.}
}
\value{
For \code{ConsortiumMetabolism} objects, returns
a tibble with species, metabolite and flux
information. For \code{ConsortiumMetabolismSet} and
\code{ConsortiumMetabolismAlignment} objects, returns
a list of such tibbles.

A list with the community data.

A named list of \code{ConsortiumMetabolism} objects.
}
\description{
Returns the consortia data in a tabular format.
}
\section{Methods (by class)}{
\itemize{
\item \code{consortia(ConsortiumMetabolism)}: Get the Community

\item \code{consortia(ConsortiumMetabolismAlignment)}: Not applicable for alignments

\item \code{consortia(ConsortiumMetabolismSet)}: Get the list of
\code{ConsortiumMetabolism} objects from a
\code{ConsortiumMetabolismSet}

}}
\examples{
cm <- synCM("test", n_species = 3, max_met = 5)
consortia(cm)

}
