% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-description.R
\name{description}
\alias{description}
\alias{description<-}
\alias{description,ConsortiumMetabolism-method}
\alias{description,ConsortiumMetabolismSet-method}
\alias{description,ConsortiumMetabolismAlignment-method}
\alias{description<-,ConsortiumMetabolism-method}
\alias{description<-,ConsortiumMetabolismSet-method}
\alias{description<-,ConsortiumMetabolismAlignment-method}
\title{Get or Set Object Description}
\usage{
description(object)

description(object) <- value

\S4method{description}{ConsortiumMetabolism}(object)

\S4method{description}{ConsortiumMetabolismSet}(object)

\S4method{description}{ConsortiumMetabolismAlignment}(object)

\S4method{description}{ConsortiumMetabolism}(object) <- value

\S4method{description}{ConsortiumMetabolismSet}(object) <- value

\S4method{description}{ConsortiumMetabolismAlignment}(object) <- value
}
\arguments{
\item{object}{A \link{ConsortiumMetabolism},
\link{ConsortiumMetabolismSet}, or
\link{ConsortiumMetabolismAlignment} object.}

\item{value}{Character scalar specifying the new
description.}
}
\value{
A character scalar (getter), or the modified
object (setter).
}
\description{
Returns or sets the description of a ramen object.
}
\examples{
cm <- synCM("test", n_species = 3, max_met = 5)
description(cm)

}
