% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.generate_bootstrap_Y_als}
\alias{.generate_bootstrap_Y_als}
\title{Generate a bootstrap outcome matrix from fitted values and residuals}
\usage{
.generate_bootstrap_Y_als(
  fitted_mat,
  resid_mat,
  bootstrap_type = c("residual", "wild", "parametric")
)
}
\arguments{
\item{fitted_mat}{Fitted value matrix.}

\item{resid_mat}{Residual matrix.}

\item{bootstrap_type}{One of "residual", "wild", or "parametric".}
}
\value{
A bootstrap outcome matrix.
}
\description{
Generate a bootstrap outcome matrix from fitted values and residuals
}
