% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{.validate_bootstrap_als_object}
\alias{.validate_bootstrap_als_object}
\title{Validate bootstrap ALS object}
\usage{
.validate_bootstrap_als_object(object)
}
\arguments{
\item{object}{An object expected to inherit from
"bootstrap_als_factorize_joint_cov".}
}
\value{
Invisibly returns TRUE if validation passes.
}
\description{
Validate bootstrap ALS object
}
