% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extract_coefficients_als}
\alias{extract_coefficients_als}
\title{Extract coefficient vector for inference}
\usage{
extract_coefficients_als(object, include_alpha = TRUE, include_beta = TRUE)
}
\arguments{
\item{object}{A fitted object returned by als_factorize_joint_cov().}

\item{include_alpha}{Logical; include row effects.}

\item{include_beta}{Logical; include column effects.}
}
\value{
A named numeric vector of coefficients.
}
\description{
Extract coefficient vector for inference
}
