% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extract_latent_matrix_als}
\alias{extract_latent_matrix_als}
\title{Extract latent interaction matrix U V^T}
\usage{
extract_latent_matrix_als(object)
}
\arguments{
\item{object}{A fitted object returned by als_factorize_joint_cov().}
}
\value{
An N x M matrix of latent interactions.
}
\description{
Extract latent interaction matrix U V^T
}
