% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-extractCluster.R
\name{extractCluster}
\alias{extractCluster}
\alias{extractCluster,ConsortiumMetabolismSet-method}
\title{Extract a Cluster}
\usage{
extractCluster(
  object,
  node_id,
  name = NA_character_,
  description = NA_character_
)

\S4method{extractCluster}{ConsortiumMetabolismSet}(
  object,
  node_id,
  name = NA_character_,
  description = NA_character_
)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismSet} object.}

\item{node_id}{Numeric scalar giving the node to be
extracted.}

\item{name}{Character scalar specifying a name for the
selection.}

\item{description}{Character scalar describing the
selection.}
}
\value{
A \code{ConsortiumMetabolismSet} object
containing the extracted cluster.
}
\description{
Extracts a cluster from a
\code{ConsortiumMetabolismSet} object.
}
\examples{
cm1 <- synCM("comm_1", n_species = 3, max_met = 5)
cm2 <- synCM("comm_2", n_species = 4, max_met = 6)
cms <- ConsortiumMetabolismSet(
    cm1, cm2, name = "test"
)
extractCluster(cms, node_id = 1)

}
