% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cms.R
\name{filterConsortia}
\alias{filterConsortia}
\alias{filterConsortia,ConsortiumMetabolismSet-method}
\title{Filter Consortia from a Set}
\usage{
filterConsortia(object, i)

\S4method{filterConsortia}{ConsortiumMetabolismSet}(object, i)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismSet} object.}

\item{i}{Integer vector, character vector of consortium
names, or logical vector.}
}
\value{
A \code{ConsortiumMetabolismSet} containing
only the selected consortia.
}
\description{
Selects a subset of consortia from a
\code{ConsortiumMetabolismSet}, returning a fully
recomputed \code{ConsortiumMetabolismSet} containing
only the selected consortia.
}
\section{Methods (by class)}{
\itemize{
\item \code{filterConsortia(ConsortiumMetabolismSet)}: Filter consortia from a
\code{ConsortiumMetabolismSet}

}}
\examples{
cm1 <- synCM("a", n_species = 3, max_met = 5)
cm2 <- synCM("b", n_species = 3, max_met = 5)
cms <- ConsortiumMetabolismSet(cm1, cm2, name = "test")
filterConsortia(cms, 1L)

}
