% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCisSNPs.R
\name{findCisSNPs}
\alias{findCisSNPs}
\title{Find cis SNPs around a set of Variable Methylated Loci (VML)}
\usage{
findCisSNPs(VML, genotype_information, distance = 1e+06)
}
\arguments{
\item{VML}{GRanges object. Must contain a metadata column named "probes",
where each element contains a vector with the probes constituting the
VML.}

\item{genotype_information}{A data frame with information about genotyped
sites of interest. It must contain the following columns: "CHROM"
(chromosome number), "POS" (Genomic basepair position of the SNP (must be an
integer), and "ID" (SNP ID). The nomenclature of CHROM must match with the
one used in the VML seqnames column (i.e., if VML uses 1, 2,
3, X, Y or Chr1, Chr2, Chr3, ChrX, ChrY, etc. as chromosome number, the
genotype_information$CHROM values must be encoded in the same way).}

\item{distance}{The distance threshold in basepairs to be used to identify
cis SNPs. Default is 1 Mb.}
}
\value{
The same VML object with new metadata columns indicating the cis SNPs
identified for each VML and the number of SNPs surrounding each VML in the
specified window
}
\description{
Identification of genotyped Single Nucleotide Polymorphisms (SNPs) close to
each VML using a distance threshold.
}
\details{
\strong{Important}: please make sure that the positions of the VML data frame and
the ones in the genotype information are from the same genome build.
}
\examples{
## Find VML in test data
VML <- RAMEN::findVML(
  methylation_data = RAMEN::test_methylation_data,
  array_manifest = "IlluminaHumanMethylationEPICv1",
  cor_threshold = 0,
  var_method = "variance",
  var_distribution = "ultrastable",
  var_threshold_percentile = 0.99,
  max_distance = 1000
)
## Find cis SNPs around VML
# Use only 5 for demonstration purposes
VML_with_cis_snps <- RAMEN::findCisSNPs(
  VML = VML$VML[1:5, ],
  genotype_information = RAMEN::test_genotype_information,
  distance = 1e6
)

}
