% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_ame_methods.R
\name{fitted.dynamic_ame}
\alias{fitted.dynamic_ame}
\title{Fitted values from a dynamic AME model}
\usage{
\method{fitted}{dynamic_ame}(object, ...)
}
\arguments{
\item{object}{A \code{dynamic_ame} object.}

\item{...}{Ignored.}
}
\value{
A list of \code{T} matrices of fitted values, named by period.
}
\description{
Fitted values from a dynamic AME model
}
