% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_estimation.R
\name{fitted_als_factorize_joint_cov}
\alias{fitted_als_factorize_joint_cov}
\title{Construct the full fitted value matrix}
\usage{
fitted_als_factorize_joint_cov(
  object,
  X_row = NULL,
  X_col = NULL,
  X_dyad = NULL
)
}
\arguments{
\item{object}{A fitted object returned by als_factorize_joint_cov().}

\item{X_row}{Optional row covariate matrix.}

\item{X_col}{Optional column covariate matrix.}

\item{X_dyad}{Optional dyad covariate array.}
}
\value{
An N x M matrix of fitted values.
}
\description{
Construct the full fitted value matrix
}
