% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getCluster.R
\name{getCluster}
\alias{getCluster}
\alias{getCluster,ConsortiumMetabolismSet-method}
\title{Get a cluster from a \code{ConsortiaMetabolismSet} Object}
\usage{
getCluster(object, node_id, name = NA_character_, description = NA_character_)

\S4method{getCluster}{ConsortiumMetabolismSet}(object, node_id, name = NA_character_, description = NA_character_)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismSet} object}

\item{node_id}{Numeric scalar giving the node to be extracted.}

\item{name}{Character scalar specifying a name for the selection.}

\item{description}{Character scalar describing the selection.}
}
\value{
A cluster from a \code{ConsortiumMetabolismSet} object
}
\description{
Gets a cluster from a \code{ConsortiumMetabolismSet} object.
}
\examples{
\donttest{
cm1 <- synCM("comm_1", n_species = 3, max_met = 5)
cm2 <- synCM("comm_2", n_species = 4, max_met = 6)
cms <- ConsortiumMetabolismSet(cm1, cm2, name = "test")
getCluster(cms, node_id = 1)
}

}
