% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getCo.R
\name{getCo}
\alias{getCo}
\alias{getCo,ConsortiumMetabolism-method}
\alias{getCo,ConsortiumMetabolismAlignment-method}
\title{Get the Consortia}
\usage{
getCo(object)

\S4method{getCo}{ConsortiumMetabolism}(object)

\S4method{getCo}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismAlignment} object.}
}
\value{
For \code{ConsortiumMetabolism} objects, returns a tb with species,
metabolite and flux information. For \code{ConsortiumMetabolismSet} and
\code{ConsortiumMetabolismAlignment} objects, returns a list of such tibbles.

A list with the community data.
}
\description{
Returns the consortia data in a tabular format.
}
\section{Methods (by class)}{
\itemize{
\item \code{getCo(ConsortiumMetabolism)}: Get the Community

\item \code{getCo(ConsortiumMetabolismAlignment)}: Not applicable for alignments

}}
\examples{
cm <- synCM("test", n_species = 3, max_met = 5)
getCo(cm)

}
