% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getEdges.R
\name{getEdges}
\alias{getEdges}
\alias{getEdges,ConsortiumMetabolism-method}
\title{Get Edges From a \code{ConsortiumMetabolism} Object}
\usage{
getEdges(object)

\S4method{getEdges}{ConsortiumMetabolism}(object)
}
\arguments{
\item{object}{a \code{ConsortiumMetabolism} object}
}
\value{
A list of edges in the community.
}
\description{
Get Edges From a \code{ConsortiumMetabolism} Object
}
\section{Methods (by class)}{
\itemize{
\item \code{getEdges(ConsortiumMetabolism)}: Get Edges From a \code{ConsortiumMetabolism} Object

}}
