% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getFunctionalGroups.R
\name{getFunctionalGroups}
\alias{getFunctionalGroups}
\alias{getFunctionalGroups,ConsortiumMetabolismSet-method}
\title{Get Functional Groups}
\usage{
getFunctionalGroups(object, ...)

\S4method{getFunctionalGroups}{ConsortiumMetabolismSet}(object)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismSet} object.}

\item{...}{Additional arguments to be passed to specific methods.}
}
\value{
A dendrogram object representing the hierarchical clustering of
species into functional groups. The plot of the dendrogram is also displayed.
}
\description{
Calculates and returns functional groups based on metabolic reactions.
For \code{ConsortiumMetabolismSet} objects, this involves analyzing shared
reactions across species to identify clusters of species with similar
metabolic capabilities.
}
\details{
This method is currently implemented for \code{ConsortiumMetabolismSet}
objects. Future versions will extend functionality to
\code{ConsortiumMetabolismAlignment} objects to allow for comparative
functional group analysis across different alignments.
}
