% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getMet.R
\name{getMet}
\alias{getMet}
\alias{getMet,ConsortiumMetabolism-method}
\alias{getMet,ConsortiumMetabolismSet-method}
\alias{getMet,ConsortiumMetabolismAlignment-method}
\title{Get Metabolites}
\usage{
getMet(object)

\S4method{getMet}{ConsortiumMetabolism}(object)

\S4method{getMet}{ConsortiumMetabolismSet}(object)

\S4method{getMet}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolism} or
\code{ConsortiumMetabolismAlignment} object}
}
\value{
A character vector containing the names of metabolites in the network
}
\description{
Retrieves the metabolites involved in the metabolic network.
}
