% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value}

\item{b}{Fallback value if a is NULL}
}
\value{
Either a or b
}
\description{
Returns the left-hand side if not NULL, otherwise the right-hand side.
}
