% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{latent_ci_bootstrap_als}
\alias{latent_ci_bootstrap_als}
\title{Extract bootstrap confidence interval matrices for latent interactions}
\usage{
latent_ci_bootstrap_als(object)
}
\arguments{
\item{object}{A bootstrap object returned by bootstrap_als_factorize_joint_cov().}
}
\value{
A list with elements conf.low and conf.high.
}
\description{
Extract bootstrap confidence interval matrices for latent interactions
}
