% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_estimation.R
\name{latent_matrix_als_factorize_joint_cov}
\alias{latent_matrix_als_factorize_joint_cov}
\title{Construct the latent interaction matrix U V^T}
\usage{
latent_matrix_als_factorize_joint_cov(object)
}
\arguments{
\item{object}{A fitted object returned by als_factorize_joint_cov().}
}
\value{
An N x M matrix of latent interactions.
}
\description{
Construct the latent interaction matrix U V^T
}
