% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{latent_mean_bootstrap_als}
\alias{latent_mean_bootstrap_als}
\title{Extract bootstrap mean latent interaction matrix}
\usage{
latent_mean_bootstrap_als(object)
}
\arguments{
\item{object}{A bootstrap object returned by bootstrap_als_factorize_joint_cov().}
}
\value{
A matrix of bootstrap means for U V^T.
}
\description{
Extract bootstrap mean latent interaction matrix
}
