% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{latent_sd_bootstrap_als}
\alias{latent_sd_bootstrap_als}
\title{Extract bootstrap standard deviation matrix for latent interactions}
\usage{
latent_sd_bootstrap_als(object)
}
\arguments{
\item{object}{A bootstrap object returned by bootstrap_als_factorize_joint_cov().}
}
\value{
A matrix of bootstrap standard deviations for U V^T.
}
\description{
Extract bootstrap standard deviation matrix for latent interactions
}
