% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmGE.R
\name{lmGE}
\alias{lmGE}
\title{Compute G/E/G+E/GxE linear models for a VMRs}
\usage{
lmGE(
  VMR,
  methylation_data,
  environmental_data,
  genotyping_data,
  covariates,
  formula_covariates
)
}
\arguments{
\item{VMR}{?a single VMR that comes from a VMR_df object. This VMR_df must have ?}

\item{methylation_data}{?A data frame where each row is a probe and each column is an individual.
The names of the individuals (i.e. colnames) must be the same as the "SubjectNumber" in the metadata
and environmental objects}

\item{environmental_data}{A data frame with one sample per row. The first column must be the "SubjectNumber",
and the second one a single environmental variable of choice.}

\item{genotyping_data}{A data frame with one sample per row. It must contain the columns "SubjectNumber"
and "genotype_encoded", which corresponds to the genotypes recoded into a count of 0, 1 or 2 (as characters)
representing the number of minor allele copies.}

\item{covariates}{A data frame with all the covariate information. It must have a "SubjectNumber" column with
the individual ID, which matches with the column in genotyping data and environmental data.}

\item{formula_covariates}{A string containing the covariates to be used in the models separated by a "+"
sign (e.g., Sex + Sample_Age_Days + counts.CD4T + counts.NK + counts.Bcell + ...)}
}
\value{
?a list with the following elements:
\itemize{
\item VMR_index
\item tagCpG
\item SNP
\item environment
\item statistics
\item winning_model
}
}
\description{
This will be coupled with apply(VMR_df, 1, FUN = g_e_modelling,...)
metadata and environmental objects must have a column named "SubjectNumber" that match the individuals
}
