% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramen-data.R
\docType{data}
\name{misosoup24}
\alias{misosoup24}
\title{Example Data from MiSoSoup 2024}
\format{
A list of 56 tibbles, each with columns:
\describe{
\item{species}{Character, species identifier.}
\item{metabolite}{Character, metabolite name (BiGG
identifiers).}
\item{flux}{Numeric, metabolic flux (negative for
consumption, positive for production).}
}
}
\value{
A list of 56 tibbles.
}
\description{
List of 56 solutions from MiSoSoup for different focal
strains in different media.
}
\examples{
data("misosoup24")
head(misosoup24[[1]])
}
