% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-modify.R
\name{modify}
\alias{modify}
\alias{modify,ConsortiumMetabolismSet-method}
\title{Modify a \code{ConsortiaMetabolismSet} Object}
\usage{
modify(object)

\S4method{modify}{ConsortiumMetabolismSet}(object)
}
\arguments{
\item{object}{An object of class \code{ConsortiumMetabolismSet}}
}
\value{
A modified \code{ConsortiumMetabolismSet} object

The modified \code{ConsortiumMetabolismSet} object.
}
\description{
Modifies a \code{ConsortiumMetabolismSet} by adding or removing
\code{ConsortiumMetabolism} objects. This allows for dynamic updating of
consortium sets for comparative analyses.
}
\section{Methods (by class)}{
\itemize{
\item \code{modify(ConsortiumMetabolismSet)}: Add or Remove Consortia to/from a
\code{ConsortiumMetaboliteSet} Object

}}
