% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-setName.R
\name{name<-}
\alias{name<-}
\alias{name<-,ConsortiumMetabolism-method}
\alias{name<-,ConsortiumMetabolismSet-method}
\alias{name<-,ConsortiumMetabolismAlignment-method}
\title{Set Object Name}
\usage{
name(object) <- value

\S4method{name}{ConsortiumMetabolism}(object) <- value

\S4method{name}{ConsortiumMetabolismSet}(object) <- value

\S4method{name}{ConsortiumMetabolismAlignment}(object) <- value
}
\arguments{
\item{object}{A \link{ConsortiumMetabolism},
\link{ConsortiumMetabolismSet}, or
\link{ConsortiumMetabolismAlignment} object.}

\item{value}{Character scalar specifying the new name.}
}
\value{
The modified object with updated name.
}
\description{
Replacement method to set the name of a ramen object.
}
\examples{
cm <- synCM("old_name", n_species = 3, max_met = 5)
name(cm) <- "new_name"

}
