% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-name.R
\name{name}
\alias{name}
\alias{name<-}
\alias{name,ConsortiumMetabolism-method}
\alias{name,ConsortiumMetabolismSet-method}
\alias{name,ConsortiumMetabolismAlignment-method}
\alias{name<-,ConsortiumMetabolism-method}
\alias{name<-,ConsortiumMetabolismSet-method}
\alias{name<-,ConsortiumMetabolismAlignment-method}
\title{Get or Set Object Name}
\usage{
name(object)

name(object) <- value

\S4method{name}{ConsortiumMetabolism}(object)

\S4method{name}{ConsortiumMetabolismSet}(object)

\S4method{name}{ConsortiumMetabolismAlignment}(object)

\S4method{name}{ConsortiumMetabolism}(object) <- value

\S4method{name}{ConsortiumMetabolismSet}(object) <- value

\S4method{name}{ConsortiumMetabolismAlignment}(object) <- value
}
\arguments{
\item{object}{A \link{ConsortiumMetabolism},
\link{ConsortiumMetabolismSet}, or
\link{ConsortiumMetabolismAlignment} object.}

\item{value}{Character scalar specifying the new name.}
}
\value{
A character scalar (getter), or the modified
object (setter).
}
\description{
Returns or sets the name of a ramen object.
}
\examples{
cm <- synCM("test", n_species = 3, max_met = 5)
name(cm)

}
