% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plot,ConsortiumMetabolismSet,ANY-method}
\alias{plot,ConsortiumMetabolismSet,ANY-method}
\title{Plot a ConsortiumMetabolismSet object}
\usage{
\S4method{plot}{ConsortiumMetabolismSet,ANY}(x, label_colours = NULL, max_nodes = 20, label_size = 4)
}
\arguments{
\item{x}{A \code{ConsortiumMetabolismSet} object.}

\item{label_colours}{Optional tibble with label and
colour columns.}

\item{max_nodes}{Maximum number of dendrogram nodes.}

\item{label_size}{Numeric label size.}
}
\value{
A \code{ggplot} object (returned invisibly).
}
\description{
Plot a ConsortiumMetabolismSet object
}
\examples{
cm1 <- synCM("comm_1", n_species = 3, max_met = 5)
cm2 <- synCM("comm_2", n_species = 4, max_met = 6)
cms <- ConsortiumMetabolismSet(cm1, cm2, name = "test")
plot(cms)
}
