% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignmentHeatmap.R
\name{plotAlignmentHeatmap}
\alias{plotAlignmentHeatmap}
\title{Plot Alignment Heatmap (Deprecated)}
\usage{
plotAlignmentHeatmap(object, top = NULL, bottom = NULL)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismAlignment} object.}

\item{top}{Numeric; top fraction filter.}

\item{bottom}{Numeric; bottom fraction filter.}
}
\value{
None. Raises an error.
}
\description{
This function is deprecated and will be reimplemented in
Phase 3.
}
