% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignmentNetwork.R
\name{plotAlignmentNetwork}
\alias{plotAlignmentNetwork}
\title{Plot Alignment Network (Deprecated)}
\usage{
plotAlignmentNetwork(object, frac)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismAlignment} object.}

\item{frac}{Numeric; minimum fraction to plot.}
}
\value{
None. Raises an error.
}
\description{
This function is deprecated and will be reimplemented in
Phase 3.
}
