% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctionalGroups.R
\name{plotFunctionalGroups}
\alias{plotFunctionalGroups}
\title{Plot Functional Groups Dendrogram}
\usage{
plotFunctionalGroups(fg, k = 4, label_size = 3, label_colours = NULL)
}
\arguments{
\item{fg}{A list as returned by
\code{\link{functionalGroups}}, containing at least a
\code{dendrogram} element.}

\item{k}{Integer scalar specifying the number of
clusters to color in the dendrogram. Default
\code{4}.}

\item{label_size}{Numeric scalar specifying the text
size for species labels. Default \code{3}.}

\item{label_colours}{If not \code{NULL}, a data frame
with columns \code{label} and \code{colour} mapping
species names to colors. When \code{NULL} (default),
all labels are drawn in black.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualizes the functional groups dendrogram computed by
\code{\link{functionalGroups}}. Branches are colored by
cluster membership and species labels are placed below
the dendrogram.
}
\examples{
cm1 <- synCM("comm_1", n_species = 3, max_met = 5)
cm2 <- synCM("comm_2", n_species = 4, max_met = 6)
cms <- ConsortiumMetabolismSet(
    cm1, cm2, name = "test"
)
fg <- functionalGroups(cms)
plotFunctionalGroups(fg, k = 2)

}
\seealso{
\code{\link{functionalGroups}} for computing
the functional groups data.
}
