% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cma.R
\name{prevalence}
\alias{prevalence}
\alias{prevalence,ConsortiumMetabolismAlignment-method}
\title{Get Edge Prevalence}
\usage{
prevalence(object)

\S4method{prevalence}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{A \link{ConsortiumMetabolismAlignment} object
of type \code{"multiple"}.}
}
\value{
A data.frame with columns \code{consumed}, \code{produced},
\code{nConsortia}, and \code{proportion}.
}
\description{
Returns edge prevalence across consortia from a
multiple alignment.
}
\section{Methods (by class)}{
\itemize{
\item \code{prevalence(ConsortiumMetabolismAlignment)}: Edge prevalence from a
multiple \link{ConsortiumMetabolismAlignment}

}}
