% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{print.bootstrap_als_factorize_joint_cov}
\alias{print.bootstrap_als_factorize_joint_cov}
\title{Print method for bootstrap_als_factorize_joint_cov objects}
\usage{
\method{print}{bootstrap_als_factorize_joint_cov}(x, ...)
}
\arguments{
\item{x}{An object of class "bootstrap_als_factorize_joint_cov".}

\item{...}{Unused.}
}
\value{
Invisibly returns x.
}
\description{
Print method for bootstrap_als_factorize_joint_cov objects
}
