% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ame.R
\name{print.bootstrap_ame}
\alias{print.bootstrap_ame}
\title{Print a combined bootstrap result}
\usage{
\method{print}{bootstrap_ame}(x, ...)
}
\arguments{
\item{x}{A \code{bootstrap_ame} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a combined bootstrap result
}
