% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_conditional.R
\name{print.bootstrap_ame_conditional}
\alias{print.bootstrap_ame_conditional}
\title{Print a conditional bootstrap result}
\usage{
\method{print}{bootstrap_ame_conditional}(x, ...)
}
\arguments{
\item{x}{A \code{bootstrap_ame_conditional} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a conditional bootstrap result
}
