% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_full.R
\name{print.bootstrap_ame_full}
\alias{print.bootstrap_ame_full}
\title{Print a full-model bootstrap result}
\usage{
\method{print}{bootstrap_ame_full}(x, ...)
}
\arguments{
\item{x}{A \code{bootstrap_ame_full} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a full-model bootstrap result
}
