% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{print.bootstrap_temporal_bipartite_als}
\alias{print.bootstrap_temporal_bipartite_als}
\title{Print method for bootstrap_temporal_bipartite_als objects}
\usage{
\method{print}{bootstrap_temporal_bipartite_als}(x, ...)
}
\arguments{
\item{x}{An object of class "bootstrap_temporal_bipartite_als".}

\item{...}{Unused.}
}
\value{
Invisibly returns x.
}
\description{
Print method for bootstrap_temporal_bipartite_als objects
}
