% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_ame_methods.R
\name{print.dynamic_ame}
\alias{print.dynamic_ame}
\title{Print a fitted dynamic AME model}
\usage{
\method{print}{dynamic_ame}(x, ...)
}
\arguments{
\item{x}{A \code{dynamic_ame} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a fitted dynamic AME model
}
