% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_dynamic_ame.R
\name{print.gibbs_dynamic_ame}
\alias{print.gibbs_dynamic_ame}
\title{Print a Gibbs sampler result}
\usage{
\method{print}{gibbs_dynamic_ame}(x, ...)
}
\arguments{
\item{x}{A \code{gibbs_dynamic_ame} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Gibbs sampler result
}
