% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{print.summary.bootstrap_als_factorize_joint_cov}
\alias{print.summary.bootstrap_als_factorize_joint_cov}
\title{Print method for summary.bootstrap_als_factorize_joint_cov objects}
\usage{
\method{print}{summary.bootstrap_als_factorize_joint_cov}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class "summary.bootstrap_als_factorize_joint_cov".}

\item{digits}{Number of digits to print for numeric summaries.}

\item{...}{Unused.}
}
\value{
Invisibly returns x.
}
\description{
Print method for summary.bootstrap_als_factorize_joint_cov objects
}
