% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_ame_methods.R
\name{print.summary.dynamic_ame}
\alias{print.summary.dynamic_ame}
\title{Print a dynamic AME summary}
\usage{
\method{print}{summary.dynamic_ame}(x, digits = 5, ...)
}
\arguments{
\item{x}{A \code{summary.dynamic_ame} object.}

\item{digits}{Number of significant digits.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a dynamic AME summary
}
