% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_estimation.R
\name{residuals_als_factorize_joint_cov}
\alias{residuals_als_factorize_joint_cov}
\title{Compute residual matrix from fitted ALS object}
\usage{
residuals_als_factorize_joint_cov(
  object,
  Y,
  X_row = NULL,
  X_col = NULL,
  X_dyad = NULL
)
}
\arguments{
\item{object}{A fitted object returned by als_factorize_joint_cov().}

\item{Y}{Outcome matrix used in estimation.}

\item{X_row}{Optional row covariate matrix.}

\item{X_col}{Optional column covariate matrix.}

\item{X_dyad}{Optional dyad covariate array.}
}
\value{
An N x M residual matrix, Y - fitted.
}
\description{
Compute residual matrix from fitted ALS object
}
