% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cma.R
\name{scores}
\alias{scores}
\alias{scores,ConsortiumMetabolismAlignment-method}
\title{Get Alignment Scores}
\usage{
scores(object)

\S4method{scores}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{A \link{ConsortiumMetabolismAlignment} object.}
}
\value{
A named list of scores.
}
\description{
Returns the scores from a
\link{ConsortiumMetabolismAlignment} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{scores(ConsortiumMetabolismAlignment)}: Scores from a
\link{ConsortiumMetabolismAlignment}

}}
