% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-setDescription.R
\name{setDesc}
\alias{setDesc}
\alias{setDesc,ConsortiumMetabolism-method}
\alias{setDesc,ConsortiumMetabolismSet-method}
\alias{setDesc,ConsortiumMetabolismAlignment-method}
\title{Set Names or description of ramen Objects}
\usage{
setDesc(object, value = NA_character_)

\S4method{setDesc}{ConsortiumMetabolism}(object, value = NA_character_)

\S4method{setDesc}{ConsortiumMetabolismSet}(object, value = NA_character_)

\S4method{setDesc}{ConsortiumMetabolismAlignment}(object, value = NA_character_)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolism},
\code{ConsortiumMetabolismSet}, or \code{ConsortiumMetabolismAlignment}
object}

\item{value}{Character scalar specifying a name.}
}
\value{
A cluster from a \code{ConsortiumMetabolismSet} object
}
\description{
Sets a new name or description of \code{ConsortiumMetabolism},
\code{ConsortiumMetabolismSet}, or \code{ConsortiumMetabolismAlignment}
objects
}
