% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cma.R
\name{sharedPathways}
\alias{sharedPathways}
\alias{sharedPathways,ConsortiumMetabolismAlignment-method}
\title{Get Shared Pathways}
\usage{
sharedPathways(object)

\S4method{sharedPathways}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{A \link{ConsortiumMetabolismAlignment} object
of type \code{"pairwise"}.}
}
\value{
A data.frame of shared pathway edges.
}
\description{
Returns pathways shared between query and reference
in a pairwise alignment.
}
\section{Methods (by class)}{
\itemize{
\item \code{sharedPathways(ConsortiumMetabolismAlignment)}: Shared pathways from a
pairwise \link{ConsortiumMetabolismAlignment}

}}
