% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-show.R
\name{show,ConsortiumMetabolism-method}
\alias{show,ConsortiumMetabolism-method}
\title{Show Method for \code{ConsortiumMetabolism} Object}
\usage{
\S4method{show}{ConsortiumMetabolism}(object)
}
\arguments{
\item{object}{An object of class \code{ConsortiumMetabolism}}
}
\value{
The object, invisibly.
}
\description{
Show Method for \code{ConsortiumMetabolism} Object
}
\examples{
cm <- synCM("test", n_species = 3, max_met = 5)
show(cm)
}
