% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-show.R
\name{show,ConsortiumMetabolismAlignment-method}
\alias{show,ConsortiumMetabolismAlignment-method}
\title{Show method for \code{ConsortiumMetabolismAlignment} Objects}
\usage{
\S4method{show}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismAlignment} object.}
}
\value{
The object, invisibly.
}
\description{
Show method for \code{ConsortiumMetabolismAlignment} Objects
}
\examples{
cm1 <- synCM("comm_1", n_species = 3, max_met = 5)
cm2 <- synCM("comm_2", n_species = 4, max_met = 6)
cma <- align(cm1, cm2)
show(cma)
}
