% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-show.R
\name{show,ConsortiumMetabolismSet-method}
\alias{show,ConsortiumMetabolismSet-method}
\title{Show Method for \code{ConsortiumMetabolismSet} Object}
\usage{
\S4method{show}{ConsortiumMetabolismSet}(object)
}
\arguments{
\item{object}{An object of class \code{ConsortiumMetabolismSet}}
}
\value{
The object, invisibly.
}
\description{
Show Method for \code{ConsortiumMetabolismSet} Object
}
\examples{
cm1 <- synCM("comm_1", n_species = 3, max_met = 5)
cm2 <- synCM("comm_2", n_species = 4, max_met = 6)
cms <- ConsortiumMetabolismSet(cm1, cm2, name = "test")
show(cms)
}
