% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cma.R
\name{similarityMatrix}
\alias{similarityMatrix}
\alias{similarityMatrix,ConsortiumMetabolismAlignment-method}
\title{Get Similarity Matrix}
\usage{
similarityMatrix(object)

\S4method{similarityMatrix}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{A \link{ConsortiumMetabolismAlignment} object
of type \code{"multiple"}.}
}
\value{
A numeric n x n matrix.
}
\description{
Returns the pairwise similarity matrix from a
multiple alignment.
}
\section{Methods (by class)}{
\itemize{
\item \code{similarityMatrix(ConsortiumMetabolismAlignment)}: Similarity matrix from a
multiple \link{ConsortiumMetabolismAlignment}

}}
