% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{summarize_bootstrap_coefficients_als}
\alias{summarize_bootstrap_coefficients_als}
\title{Summarize bootstrap coefficient draws}
\usage{
summarize_bootstrap_coefficients_als(coef_draws, coef_hat, conf_level = 0.95)
}
\arguments{
\item{coef_draws}{A B x P matrix of bootstrap coefficient draws.}

\item{coef_hat}{Named numeric vector of original estimates.}

\item{conf_level}{Confidence level.}
}
\value{
A data.frame with bootstrap summaries.
}
\description{
Summarize bootstrap coefficient draws
}
