% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{summarize_bootstrap_latent_als}
\alias{summarize_bootstrap_latent_als}
\title{Summarize bootstrap latent interaction draws}
\usage{
summarize_bootstrap_latent_als(latent_draws, conf_level = 0.95)
}
\arguments{
\item{latent_draws}{A list of bootstrap latent interaction matrices.}

\item{conf_level}{Confidence level.}
}
\value{
A list of matrices: mean, sd, conf.low, conf.high, sign.prob.
}
\description{
Summarize bootstrap latent interaction draws
}
