% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.bootstrap_als_factorize_joint_cov}
\alias{summary.bootstrap_als_factorize_joint_cov}
\title{Summary method for bootstrap_als_factorize_joint_cov objects}
\usage{
\method{summary}{bootstrap_als_factorize_joint_cov}(object, ...)
}
\arguments{
\item{object}{An object of class "bootstrap_als_factorize_joint_cov".}

\item{...}{Unused.}
}
\value{
An object of class "summary.bootstrap_als_factorize_joint_cov".
}
\description{
Summary method for bootstrap_als_factorize_joint_cov objects
}
