% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_methods.R
\name{summary.bootstrap_temporal_bipartite_als}
\alias{summary.bootstrap_temporal_bipartite_als}
\title{Summary method for bootstrap_temporal_bipartite_als objects}
\usage{
\method{summary}{bootstrap_temporal_bipartite_als}(object, ...)
}
\arguments{
\item{object}{An object of class "bootstrap_temporal_bipartite_als".}

\item{...}{Unused.}
}
\value{
An object of class "summary.bootstrap_temporal_bipartite_als".
}
\description{
Summary method for bootstrap_temporal_bipartite_als objects
}
