% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_init.R
\name{svd_init}
\alias{svd_init}
\title{SVD initialization for latent factors}
\usage{
svd_init(Y, K = 2)
}
\arguments{
\item{Y}{Numeric matrix.}

\item{K}{Latent dimension.}
}
\value{
A list with U and V.
}
\description{
SVD initialization for latent factors
}
