% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_covariates.R
\docType{data}
\name{test_covariates}
\alias{test_covariates}
\title{Concomitant variable example data set}
\format{
\subsection{\code{test_covariates}}{

A data frame with 30 rows and 1 column:
\describe{
\item{\emph{rownames}}{Individual IDs}
\item{covar1}{Concomitant variable drawn from a normal distribution with
mean = 0 and sd = 1}
}
}
}
\usage{
test_covariates
}
\description{
Concomitant variable drawn from a normal distribution
}
\examples{
summary(test_covariates)
}
\keyword{datasets}
