% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_environmental_matrix.R
\docType{data}
\name{test_environmental_matrix}
\alias{test_environmental_matrix}
\title{Environmental exposures example data set}
\format{
\subsection{\code{test_environmental_matrix}}{

A data frame with 30 rows and 100 columns:
\describe{
\item{\emph{rownames}}{Individual ID}
\item{E1:100}{100 example environmental exposures; column names correspond
to exposure IDs}
}
}
}
\usage{
test_environmental_matrix
}
\description{
One hundred environmental exposure variables drawn from a normal distribution
with mean=0 and sd=1 for 30 individuals.
}
\examples{
summary(test_environmental_matrix)

}
\keyword{datasets}
