% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_genotype_information.R
\docType{data}
\name{test_genotype_information}
\alias{test_genotype_information}
\title{Genotype metadata example data set}
\format{
\subsection{\code{test_genotype_information}}{

A data frame with 8539 rows and 3 columns:
\describe{
\item{CHROM}{Chromosome}
\item{POS}{Probe genomic position (h19)}
\item{ID}{SNP ID}
}
}
}
\usage{
test_genotype_information
}
\description{
Genotype chromosome 21 position metadata obtained from a private genotyping
data set.
}
\examples{
summary(test_genotype_information)

}
\keyword{datasets}
