% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cma.R
\name{uniquePathways}
\alias{uniquePathways}
\alias{uniquePathways,ConsortiumMetabolismAlignment-method}
\title{Get Unique Pathways}
\usage{
uniquePathways(object)

\S4method{uniquePathways}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{A \link{ConsortiumMetabolismAlignment} object
of type \code{"pairwise"}.}
}
\value{
A list with \code{query} and \code{reference} data.frames.
}
\description{
Returns pathways unique to query and reference in a
pairwise alignment.
}
\section{Methods (by class)}{
\itemize{
\item \code{uniquePathways(ConsortiumMetabolismAlignment)}: Unique pathways from a
pairwise \link{ConsortiumMetabolismAlignment}

}}
\examples{
cm1 <- synCM("comm_1", n_species = 3, max_met = 5)
cm2 <- synCM("comm_2", n_species = 4, max_met = 6)
cma <- align(cm1, cm2)
uniquePathways(cma)

}
