% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-align.R
\name{align}
\alias{align}
\alias{align,ConsortiumMetabolismSet-method}
\title{Align a \code{ConsortiaMetabolismSet} Object}
\usage{
align(object)

\S4method{align}{ConsortiumMetabolismSet}(object)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismSet} object.}

\item{name}{Character scalar giving name of the alignment, if \code{NULL} inherits
from the \code{ConsortiumMetabolismSet} object.}
}
\value{
A \code{ConsortiumMetabolismAlignment} object.
}
\description{
A short description...
}
\section{Methods (by class)}{
\itemize{
\item \code{align(ConsortiumMetabolismSet)}: Align a \code{ConsortiumMetabolismSet} Object

}}
