% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-cluster.R
\name{cluster}
\alias{cluster}
\alias{cluster,ConsortiumMetabolismSet-method}
\title{Cluster a \code{ConsortiaMetabolismSet} Object}
\usage{
cluster(object)

\S4method{cluster}{ConsortiumMetabolismSet}(object)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismSet} object to modify}
}
\value{
A modified \code{ConsortiumMetabolismSet} object
}
\description{
Modifies a \code{ConsortiumMetabolismSet} by clustering the binary matrices
and storing the result to the object.
}
