% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-compareSpecies.R
\name{compareSpecies}
\alias{compareSpecies}
\alias{compareSpecies,ConsortiumMetabolismSet-method}
\title{Compare Species}
\usage{
compareSpecies(object, species)

\S4method{compareSpecies}{ConsortiumMetabolismSet}(object, species)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismSet} object.}

\item{species}{A character vector of species names to compare}
}
\value{
A list of tibbles.
}
\description{
Compares the metabolisms of two species and outputs a list of tibbles,
containing the tibbles intersection, unique, and consistency.
As the name suggests, intersection, unique contain the intersection and
unique pathways per species compared, while the consistency tibble contains
information on whether or not a specie's set of pathways is consistent in
all of the consortia in which it is present or not. If all species contain
the same edges in all consortia in which they appear, this tibble will be
returned with 0 rows.
For \code{ConsortiumMetabolismSet} objects.
}
\details{
This method is currently implemented for \code{ConsortiumMetabolismSet}
objects. Future versions will extend functionality to
\code{ConsortiumMetabolism} objects to allow for the analysis of species
within a single consortium different alignments.
}
\section{Methods (by class)}{
\itemize{
\item \code{compareSpecies(ConsortiumMetabolismSet)}: Compare Species in a ConsortiumMetabolismSet

}}
