% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cma.R,
%   R/methods-accessors-cms.R
\name{consortia}
\alias{consortia}
\alias{consortia,ConsortiumMetabolismAlignment-method}
\alias{consortia,ConsortiumMetabolismSet-method}
\title{Get the Constituent Consortia}
\usage{
consortia(object)

\S4method{consortia}{ConsortiumMetabolismAlignment}(object)

\S4method{consortia}{ConsortiumMetabolismSet}(object)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismSet} object.}
}
\value{
A named list of \code{ConsortiumMetabolism}
objects.
}
\description{
Returns the \code{ConsortiumMetabolism} objects that a
container is built from. Defined for
\code{ConsortiumMetabolismSet} (the consortia in the set)
and \code{ConsortiumMetabolismAlignment} (the consortia
that produced the alignment). The plural noun reflects
that the result is always a collection; for a single
\code{ConsortiumMetabolism}, use
\code{\link[=as.data.frame]{as.data.frame()}} to obtain
the underlying edge list.
}
\section{Methods (by class)}{
\itemize{
\item \code{consortia(ConsortiumMetabolismAlignment)}: Not applicable to
\code{ConsortiumMetabolismAlignment}. By design the
class is a lightweight result object: it records the
names of its inputs (\code{QueryName} /
\code{ReferenceName}) but does not retain copies of
the \code{ConsortiumMetabolism} objects themselves.
Look up the originating CMs by name from the source
\code{ConsortiumMetabolismSet}.

\item \code{consortia(ConsortiumMetabolismSet)}: Get the list of
\code{ConsortiumMetabolism} objects from a
\code{ConsortiumMetabolismSet}.

}}
\examples{
cm1 <- synCM("a", n_species = 3, max_met = 5)
cm2 <- synCM("b", n_species = 3, max_met = 5)
cms <- ConsortiumMetabolismSet(list(cm1, cm2), name = "demo")
consortia(cms)

}
