% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-consortia.R
\name{consortia}
\alias{consortia}
\alias{consortia,ConsortiumMetabolism-method}
\alias{consortia,ConsortiumMetabolismAlignment-method}
\title{Get the Consortia}
\usage{
consortia(object)

\S4method{consortia}{ConsortiumMetabolism}(object)

\S4method{consortia}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismAlignment} object.}
}
\value{
For \code{ConsortiumMetabolism} objects, returns
a tibble with species, metabolite and flux
information. For \code{ConsortiumMetabolismSet} and
\code{ConsortiumMetabolismAlignment} objects, returns
a list of such tibbles.

A list with the community data.
}
\description{
Returns the consortia data in a tabular format.
}
\section{Methods (by class)}{
\itemize{
\item \code{consortia(ConsortiumMetabolism)}: Get the Community

\item \code{consortia(ConsortiumMetabolismAlignment)}: Not applicable for alignments

}}
\examples{
cm <- synCM("test", n_species = 3, max_met = 5)
consortia(cm)

}
