% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsortiumMetabolismSet.R
\name{ConsortiumMetabolismSet}
\alias{ConsortiumMetabolismSet}
\title{Set of \code{ConsortiumMetabolism} Objects}
\usage{
ConsortiumMetabolismSet(..., name = NA_character_)
}
\arguments{
\item{...}{Lists or individual \code{ConsortiumMetabolism} objects.}

\item{name}{Character scalar giving the name of the set.}
}
\description{
Set of \code{ConsortiumMetabolism} Objects
}
