% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCisSNPs.R
\name{findCisSNPs}
\alias{findCisSNPs}
\title{Find cis SNPs around a set of Variable Methylated Regions}
\usage{
findCisSNPs(VMRs_df, genotype_information, distance = 1e+06)
}
\arguments{
\item{VMRs_df}{A GRanges object converted to a data frame. Must contain the following columns:
"seqnames", "start", "end". These columns are present automatically when doing the object conversion.}

\item{genotype_information}{A data frame with information about genotyped sites of interest. It must contain the following
columns: "Chr" - chromosome number, "Position" - Genomic basepair position of SNP in the corresponding
chromosome (must contain values of class int), and "Name" - SNP ID.}

\item{distance}{The distance threshold to be used to identify cis SNPs. Default is 1 Mb}
}
\value{
A VMR_df: a VMR_df object with a new column including the cis SNPs identified for each VMR, and the number
of SNPs surrounding each VMR in the specified window
}
\description{
Identification of genotyped SNPs close to each VMR using a distance threshold. \strong{Important}: please make sure that the
positions of the VMR data frame and the ones in the genotype information are from the same genome build.
}
