% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCisSNPs.R
\name{findCisSNPs}
\alias{findCisSNPs}
\title{Find cis SNPs around a set of VMRs}
\usage{
findCisSNPs(VMRs_df, genotyping_probes, distance = 1e+06)
}
\arguments{
\item{VMRs_df}{A data frame converted from a Genomic Ranges object. Must contain the following columns:
"probes", containing a list where each element contains a vector with the probes constituting a  VMR}

\item{genotyping_probes}{?A data frame suitable to be converted to a Genomic Ranges object,
containing the genomic coordinates of genotyped sites}

\item{distance}{The distance threshold to be used to identify cis SNPs}
}
\value{
a list with the number of SNPs per VMR and a VMR_df with a new column including the cis SNPs
identified for each VMR.
}
\description{
This function identifies genotyped SNPs close to each VMR using a distance threshold.
}
