% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCisSNPs.R
\name{findCisSNPs}
\alias{findCisSNPs}
\title{Find cis SNPs around a set of VMRs}
\usage{
findCisSNPs(VMRs_df, genotype_information, distance = 1e+06)
}
\arguments{
\item{VMRs_df}{A data frame converted from a Genomic Ranges object. Must contain the following columns:
"seqnames", "start", "end".}

\item{genotype_information}{A data frame with information about genotyped sites of interest. It must contain the following
columns: "Chr" - Number of chromosome, "Position" - Genomic position of the chromosome (must contain values of class int),
and "Name" - ID of the SNP.}

\item{distance}{The distance threshold to be used to identify cis SNPs. Default is 1 Mb}
}
\value{
a VMR_df: a VMR_df object but now with a new column including the cis SNPs identified for each VMR, and the number
of SNPs surrounding each VMR in the specified window
}
\description{
Identification of genotyped SNPs close to each VMR using a distance threshold. Important: please make sure that the
positions of the VMR data frame and the ones in the genotype information are from the same genome build.
}
