% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVMRs.R
\name{findVMRs}
\alias{findVMRs}
\title{Identify Variable Methylated Regions in microarrays}
\usage{
findVMRs(
  array_manifest,
  methylation_data,
  cor_threshold,
  var_method,
  var_threshold_percentile = 0.9,
  max_distance = 1000
)
}
\arguments{
\item{array_manifest}{Information about the probes on the array. Requires the columns MAPINFO (basepair position
of the probe in the genome), CHR (chromosome), TargetID(probe name) and STRAND (this is very important to set up, since
the VMRs will only be created based on CpGs on the same strand; if the positions are reported based on a single DNA strand,
this should contain either a vector of only "+", "-" or "*" for all of the probes).}

\item{methylation_data}{A data frame containing M or B values, with samples as columns and probes as rows}

\item{cor_threshold}{Numeric value (0-1) to be used as the median correlation threshold for identifying VMRs (i.e.
all VMRs will have a median pairwise probe correlation of this parameter).}

\item{var_method}{Method to use to measure variability in the data set. The options are "mad" (median absolute deviation)
or "variance".}

\item{var_threshold_percentile}{The percentile to be used as cut off to define Highly Variable Probes (and
therefore VMRs). The default is 0.9 because this percentile has be traditionally used in previous studies.}

\item{max_distance}{Maximum distance allowed for two probes to me grouped into a region. The default is 1000
because this percentile has be traditionally used in previous studies.}
}
\value{
A list with the following elements:
\itemize{
\item $var_score_threshold: Threshold used to define Highly Variable Probes (mad or variance, depending on the specified choice).
\item $highly_variable_probes a data frame with the probes that passed the variability score threshold imposed by the user, and their variability score (MAD score or variance).
\item $canonical_VMRs: a GRanges object with strict candidate VMRs - regions composed of 2 or more
contiguous, correlated and proximal Highly Variable Probes; thresholds depend on the ones specified
by the user)
\item $non_canonical_VMRs: a GRanges object with highly variable probes without neighboring
CpGs measured in \emph{max_distance} on the array.
}
}
\description{
Identifies Highly Variable Probes (HVP) and merge them into Variable Methylated Regions (VMRs),
given a sorted illumina manifest. Note: this function uses GenomicRanges::reduce() to group the
regions, which is strand-sensitive. In the illumina microarrays, the MAPINFO for all the probes
is usually provided as for the + strand. If you are using this array, we recommend to previously
convert the strand of all the probes to "+". OPTIONAL: This function supports parallel computing for increased speed. To do so, you have to set the parallel backend
in your R session BEFORE running the function (e.g., doFuture::registerDoFuture()) and then the evaluation strategy (e.g., future::plan(multisession)). After that,
the function can be run as usual.
}
