% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-functionalGroups.R
\name{functionalGroups}
\alias{functionalGroups}
\alias{functionalGroups,ConsortiumMetabolismSet-method}
\title{Get Functional Groups}
\usage{
functionalGroups(object, k = 4, ...)

\S4method{functionalGroups}{ConsortiumMetabolismSet}(object, k = 4, label_size = 6, label_colours = NULL)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismSet} object.}

\item{k}{An integer scalar specifying the number of
clusters to color in the dendrogram.}

\item{...}{Additional arguments to be passed to
specific methods.}

\item{label_size}{Numeric scalar specifying label size
in the output plot.}

\item{label_colours}{If not \code{NULL}, a tibble with
columns \code{label} and \code{colour}.}
}
\value{
A list (returned invisibly) containing:
\itemize{
\item plot: The ggplot2 dendrogram visualization
\item dendrogram: The dendrogram object
\item similarity_matrix: Matrix of Jaccard
similarities between species
\item species_combinations: Tibble with pairwise
species comparisons
\item reactions_per_species: Tibble mapping species
to their reactions
}
The plot is automatically displayed.
}
\description{
Calculates and returns functional groups based on
metabolic reactions. For
\code{ConsortiumMetabolismSet} objects, this involves
analyzing shared reactions across species to identify
clusters of species with similar metabolic
capabilities.
}
\details{
This method is currently implemented for
\code{ConsortiumMetabolismSet} objects. Future versions
will extend functionality to
\code{ConsortiumMetabolismAlignment} objects to allow
for comparative functional group analysis across
different alignments.
}
\examples{
\donttest{
cm1 <- synCM("comm_1", n_species = 3, max_met = 5)
cm2 <- synCM("comm_2", n_species = 4, max_met = 6)
cms <- ConsortiumMetabolismSet(
    cm1, cm2, name = "test"
)
functionalGroups(cms, k = 2)
}

}
