% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getCo.R
\name{getCo}
\alias{getCo}
\alias{getCo,ConsortiumMetabolism-method}
\title{Get the Consortia}
\usage{
getCo(object)

\S4method{getCo}{ConsortiumMetabolism}(object)
}
\arguments{
\item{object}{An object of class \code{ConsortiumMetabolism}}
}
\value{
For \code{ConsortiumMetabolism} objects, returns a tb with species,
metabolite and flux information. For \code{ConsortiumMetabolismSet} and
\code{ConsortiumMetabolismAlignment} objects, returns a list of such tibbles.

A list with the community data.
}
\description{
Returns the consortia data in a tabular format.
}
\section{Methods (by class)}{
\itemize{
\item \code{getCo(ConsortiumMetabolism)}: Get the Community

}}
