% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getCo.R
\name{getCo}
\alias{getCo}
\alias{getCo,ConsortiumMetabolism-method}
\title{Get the Consortia}
\usage{
getCo(object)

\S4method{getCo}{ConsortiumMetabolism}(object)
}
\arguments{
\item{object}{An object of class \code{ConsortiumMetabolism}}
}
\value{
A list with the community data.
}
\description{
Returns the consortia of a single \code{ConsortiumMetabolism} object in a
\code{tibble} format or a list of consortias in \code{tibble} format for
\code{ConsortiumMetabolismSet} and \code{ConsortiumMetabolismAlignment}
objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{getCo(ConsortiumMetabolism)}: Get the Community

}}
