% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getEdges.R
\name{getEdges}
\alias{getEdges}
\alias{getEdges,ConsortiumMetabolism-method}
\title{Get Edges From a \code{ConsortiumMetabolism} Object}
\usage{
getEdges(object)

\S4method{getEdges}{ConsortiumMetabolism}(object)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolism} object}
}
\value{
A tibble containing edge information including:
\itemize{
\item consumed/produced metabolites
\item number of species involved
\item consumption/production sums
\item effective consumption/production metrics
}
}
\description{
Retrieves the edges representing metabolic interactions between species.
}
\section{Methods (by class)}{
\itemize{
\item \code{getEdges(ConsortiumMetabolism)}: Get Edges From a \code{ConsortiumMetabolism} Object

}}
