% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getEdges.R
\name{getEdges}
\alias{getEdges}
\alias{getEdges,ConsortiumMetabolism-method}
\alias{getEdges,ConsortiumMetabolismSet-method}
\title{Get Edges From a \code{ConsortiumMetabolism} Object}
\usage{
getEdges(object, ...)

\S4method{getEdges}{ConsortiumMetabolism}(object)

\S4method{getEdges}{ConsortiumMetabolismSet}(
  object,
  type = c("all", "pan-cons", "niche", "core", "aux"),
  quantileCutoff = 0.1
)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolism} or
\code{ConsortiumMetabolismSet} object}

\item{...}{Object-specific arguments. See methods for details.}

\item{type}{Character scalar giving the type of edges to output.}

\item{quantileCutoff}{Numeric scalar between 0 and 1 giving the quantile
threshold to use for filtering edges. For "pan-cons" and "core" types,
edges above \code{1 - quantileCutoff} are returned. For "niche" and "aux"
types, edges below \code{quantileCutoff} are returned. Defaults to 0.1
(i.e., top/bottom 10\\%).}
}
\value{
A tibble containing edge information including:
\itemize{
\item consumed/produced metabolites
\item number of species involved
\item consumption/production sums
\item effective consumption/production metrics
}
}
\description{
Retrieves the edges representing metabolic interactions between species.
The argument \code{type} can be used to return only specific types of edges
from a \code{ConsortiumMetabolismSet} object.
}
\details{
For \code{ConsortiumMetabolismSet} objects, the \code{type} parameter
determines which edges are returned:
\itemize{
\item \code{all} - Returns all edges
\item \code{pan-cons} - Returns edges present in most consortia (top
\code{1 - quantileCutoff} fraction)
\item \code{niche} - Returns rare edges (bottom \code{quantileCutoff}
fraction of consortia)
\item \code{core} - Returns edges involving many species (top
\code{1 - quantileCutoff} fraction)
\item \code{aux} - Returns edges involving few species (bottom
\code{quantileCutoff} fraction)
}
}
\section{Methods (by class)}{
\itemize{
\item \code{getEdges(ConsortiumMetabolism)}: Get Edges From a \code{ConsortiumMetabolism} Object

\item \code{getEdges(ConsortiumMetabolismSet)}: Get Edges From a \code{ConsortiumMetabolismSet} Object

}}
