% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-getEdges.R
\name{getEdges}
\alias{getEdges}
\alias{getEdges,ConsortiumMetabolism-method}
\alias{getEdges,ConsortiumMetabolismSet-method}
\title{Get Edges From a \code{ConsortiumMetabolism} Object}
\usage{
getEdges(object, ...)

\S4method{getEdges}{ConsortiumMetabolism}(object)

\S4method{getEdges}{ConsortiumMetabolismSet}(
  object,
  type = c("all", "pan-cons", "niche", "core", "aux"),
  perc = 0.1
)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolism} object}

\item{...}{Object specific arguments.}

\item{type}{Character scalar giving the type of edges to output.}

\item{perc}{Numeric scalar giving the percentage to use for the quantile
calculation.}
}
\value{
A tibble containing edge information including:
\itemize{
\item consumed/produced metabolites
\item number of species involved
\item consumption/production sums
\item effective consumption/production metrics
}
}
\description{
Retrieves the edges representing metabolic interactions between species.
The argument \code{type} can be used to return only specfic types of Edges from
a \code{ConsortiumMetabolismSet} type object.
\itemize{
\item \code{all} will return all edges
\item \code{pan-cons} will return only edges that exist in all consortia
\item \code{niche} will return niche consortia. A niche is defined a
}
}
\section{Methods (by class)}{
\itemize{
\item \code{getEdges(ConsortiumMetabolism)}: Get Edges From a \code{ConsortiumMetabolism} Object

\item \code{getEdges(ConsortiumMetabolismSet)}: Get Edges From a \code{ConsortiumMetabolismSet} Object

}}
