% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmGE.R
\name{lmGE}
\alias{lmGE}
\title{Selection of the best VMR explanatory G, E, G+E or GxE model}
\usage{
lmGE(
  selected_variables,
  summarized_methyl_VMR,
  genotype_matrix,
  environmental_matrix,
  covariates = NULL,
  model_selection = "AIC"
)
}
\arguments{
\item{selected_variables}{A data frame obtained with RAMEN::selectVariables(). This data frame must contain three columns: 'VMR_index' with characters of an unique ID of each VMR; ´selected_genot' and 'selected_env' with the SNPs and environmental variables, respectively, that will be used for fitting the genotype (G), environment (E), additive (G + E) or interaction (G x E) models. The columns 'selected_env' and 'selected_genot' must contain lists as elements; VMRs with no environmental or genotype selected variables must contain an empty list with NULL, NA or "" inside.}

\item{summarized_methyl_VMR}{A data frame containing each individual's VMR summarized region methylation. It is suggested to use the output of RAMEN::summarizeVMRs(). It must have individuals as rows, and VMRs as columns. The names of the columns must correspond to the index of said VMR, and it must match the index of VMR_df$VMR_index. The names of the ROWS must be the sample IDs, and must match with the IDs of the other matrices.}

\item{genotype_matrix}{A matrix of number-encoded genotypes. Each column should correspond to a sample, and each row to a SNP site. The suggested encoding is 2 (AA), 1 (AB) and 0 (BB). The COLUMN names must correspond with individual IDs.}

\item{environmental_matrix}{A matrix of environmental variables. Only numeric values are supported. In case of having factors, it is recommended to encode them as numbers or re-code them into dummy variables if there are more than 2 levels.Each column should correspond to an environmental variable and each row to an individual. The ROW names must be the individual IDs.}

\item{covariates}{A matrix containing the variables that will be adjusted for in the final GxE models (e.g. cell type proportions, age, etc.)-.Each column should correspond to a covariate and each row to an individual. The ROW names must be the individual IDs.}

\item{model_selection}{Which metric to use to select the best model for each VMR. Supported options are "AIC" or BIC". More information about which one to use can be found in the Details section.}
}
\value{
A data frame with the following columns:
\itemize{
\item VMR_index: The unique ID of the VMR
\item model_group: The group to which the winning model belongs to (i.e., G, E, G+E or GxE)
\item variables: The variable(s) that are present in the winning model (excluding the covariates, which are included in all the models)
\item F_val: F value from the ANOVA test comparing the basal model (i.e., only covariates) to the full model (i.e., including the G/E/G+E/G+E+GxE variables of the winning model respectively)
\item p_val: p value from the ANOVA testing whether the more complex/full model is significantly better at capturing the data than the simpler/basal model (i.e., only the covariates) . If the resulting p-value is sufficiently low (usually less than 0.05), we can interpret that the more complex model is significantly better than the basal model, and thus favor the G/E/G+E/GxE model. If the p-value is not sufficiently low (usually greater than 0.05), we should favor the basal model.
\item FDR: False Discovery Rate applied to the ANOVA p values of the winning models. This controls the expected proportion of rejected null hypotheses that are false (incorrect rejections of the null) in a multiple hypothesis testing setting (i.e., one test per VMR).
\item tot_r_squared: R squared of the winning model
\item g_r_squared: Estimated R2 allocated to the G in the winning model, if applicable.
\item e_r_squared: Estimated R2 allocated to the E in the winning model, if applicable.
\item gxe_r_squared: Estimated R2 allocated to the interaction in the winning model (GxE), if applicable.
\item AIC/BIC: Model comparison value used to select the best model in each VMR (depending on the option specified in the argument model_selection).
\item second_winner: The second group that possess the next best model after the winning one (i.e., G, E, G+E or GxE). This column may have NA if the variables selected in selected variables correspond only to one group (G or E), so that there is no different group to compare to.
\item delta_aic/delta_bic: The difference of AIC or BIC value (depending on the option specified in the argument model_selection) of the winning model and the best model from the second_winner group (i.e., G, E, G+E or GxE). This column may have NA if the variables selected in selected variables correspond only to one group (G or E), so that there is no different group to compare to.
\item delta_r_squared: The R2 of the winning model - R2 of the second winner model.This column may have NA if the variables selected in selected variables correspond only to one group (G or E), so that there is no different group to compare to.
}
}
\description{
For each Variable Methylated Region (VMR), this function fits a set of genotype (G), environment (E), pairwise additive (G + E) or pairwise interaction (G x E) models, one variable at a time, and selects the best fitting one. Additional information for each winning model is provided, such as its R2, its significance from an F test comparing it to a basal model (i.e., only the covariates), the delta AIC/BIC to the next best model from a different category, and the explained variance decomposed for the G, E and GxE components (when applicable).
}
\details{
For each VMR, this function computes a set of models using the variables indicated in the selected_variables object. From the indicated G and E variables, lmGE() fits four groups of models:
\itemize{
\item G: Genetics model - fitted one SNP at a time.
\item E: Environmental model - fitted one environmental variable at a time.
\item G+E: Additive model - fitted for each pairwise combination of G and E variables indicated in selected_variables.
\item GxE: Interaction model - fitted for each pairwise combination of G and E variables indicated in selected_variables.
}

\strong{Model selection}

Following the model fitting stage, the best model PER GROUP is selected using Akaike Information Criterion (AIC) or Bayesian Information Criteron (BIC). Both of these metrics are statistical approaches to select the right model in the same data set. Both of these metrics have strengths and limitations that make them excel in different situations.
Even though BIC is more useful in selecting a correct model, which is the goal of this tool, rather than finding the best model for predicting future observations, where the AIC is more appropriate, we recommend the use of AIC. This is because BIC assumes that the true model is in the set of compared models. Since we are fitting models with individual variables, and we assume that DNAme variability is more likely to be influenced by more than one single SNP/environmental exposure at a time, we hypothetize that in most cases, the true model will not be in the set of compared models. On the other hand, AIC excels in those problems where all models in the model space are incorrect or there is at most one correct model in the model space, and AIC is preferentially used in cases where f(x) is unknown and our selected model could belong to a very large class of functions since the relationship could be pretty complex. However, both metrics tend to pick the same model in a large number of scenarios. We suggest the users to read Arijit Chakrabarti and Jayanta K. Ghosh (2011) for further information about the difference between these metrics.

After selecting the best model per group (G,E,G+E pr GxE), the model with the lowest AIC or BIC is declared as the winning model. The delta AIC/BIC and difference of R2 is computed relative to the model with the second lowest AIC/BIC (i.e., the best model from a different group to the winning one), and reported in the final object.

\strong{Analysis of variance and variance decomposition}

After defining the best explanatory model for each VMR, an ANOVA test is conducted to test whether this model is significantly better at explaining the DNAme variability of the respective region than the basal model (i.e., the model including only the specified covariates). The respective F and p value is reported in the final returned object. Additionally, a False Discovery Rate method is applied to the p values to address the multiple hypothesis testing.
Finally, the variance is decomposed and the relative R2 contribution of each of the variables of interest (G, E and GxE) is reported. This decomposition is done using the relaimpo R package, using the Lindeman, Merenda and Gold (lmg) method, which is based on the heuristic approach of averaging the relative R contribution of each variable over all input orders in the linear model. For further information, we suggest the users to read the documentation and publication of the relaimpo R package.

Disclaimer: When the winning model is an interaction one, the relative R2 contribution is estimated keeping the covariates always in the model (i.e., they do not change orders); this can lead to an underestimation of the G,E and GxE relative contribution, but estimating these in interaction models is very computationally intensive in the presence of a medium to high number of covariates in the model. Therefore, since this package is designed to handle genome-wide data sets, we favor speed in this scenario.
}
