% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medCorVMR.R
\name{medCorVMR}
\alias{medCorVMR}
\title{Compute the median correlation of a Genomic Ranges data frame}
\usage{
medCorVMR(VMR_df, data_methylation)
}
\arguments{
\item{VMR_df}{A data frame converted from a Genomic Ranges object. Must contain the following columns: "probes", containing a list where each
element contains a vector with the probes constituting a  VMR}

\item{data_methylation}{A data frame containing M or B values, with samples as columns and probes as rows}
}
\value{
A data frame like VMR_df, but with an extra column per region containing the median pairwise correlation
}
\description{
Given a set of candidate VMRs, this function will compute the pairwise correlation of the probes
in each VMR and report its median.
}
