% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medCorVMR.R
\name{medCorVMR}
\alias{medCorVMR}
\title{Compute the median probe methylation pearson correlation for each Variable Methylated Region (VMR).}
\usage{
medCorVMR(VMR_df, methylation_data)
}
\arguments{
\item{methylation_data}{A data frame containing M or B values, with samples as columns and probes as rows. Data is expected to have already passed through quality control and cleaning steps.}
}
\value{
A data frame like VMR_df with an extra column per region containing the median pairwise correlation.
}
\description{
This function will take a GRanges object converted into a data frame, where each row corresponds to a
Variable Methylated Region. Then, it computes the pairwise correlation of the probes of each VMR and reports
its median pairwise probe correlation.
}
\details{
This function supports parallel computing for increased speed. To do so, you have to set the parallel backend
in your R session before running the function (e.g., doFuture::registerDoFuture()) and then the evaluation strategy (e.g., future::plan(multisession)). After that,
the function can be run as usual. It is recommended to also set options(future.globals.maxSize= +Inf).
}
