% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cm.R,
%   R/methods-accessors-cma.R, R/methods-accessors-cms.R
\name{metabolites}
\alias{metabolites}
\alias{metabolites,ConsortiumMetabolism-method}
\alias{metabolites,ConsortiumMetabolismAlignment-method}
\alias{metabolites,ConsortiumMetabolismSet-method}
\title{Get Metabolites}
\usage{
metabolites(object)

\S4method{metabolites}{ConsortiumMetabolism}(object)

\S4method{metabolites}{ConsortiumMetabolismAlignment}(object)

\S4method{metabolites}{ConsortiumMetabolismSet}(object)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolism} or
\code{ConsortiumMetabolismAlignment} object.}
}
\value{
A character vector containing the names of
metabolites in the network.
}
\description{
Retrieves the metabolites involved in the metabolic
network.
}
\examples{
cm <- synCM("test", n_species = 3, max_met = 5)
metabolites(cm)

}
