% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cms.R
\name{overlapMatrix}
\alias{overlapMatrix}
\alias{overlapMatrix,ConsortiumMetabolismSet-method}
\title{Get Overlap Matrix}
\usage{
overlapMatrix(object)

\S4method{overlapMatrix}{ConsortiumMetabolismSet}(object)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismSet} object.}
}
\value{
A numeric \eqn{n \times n} matrix of pairwise
dissimilarities, where \eqn{n} is the number of
consortia. Row and column names are consortium
names.
}
\description{
Returns the pairwise dissimilarity matrix from a
\code{ConsortiumMetabolismSet} object. Values are
\code{1 - FOS} (Functional Overlap Score), so 0
indicates identical consortia and 1 indicates no
shared pathways.
}
\examples{
cm1 <- synCM("comm_1", n_species = 3, max_met = 5)
cm2 <- synCM("comm_2", n_species = 4, max_met = 6)
cms <- ConsortiumMetabolismSet(
    cm1, cm2, name = "test"
)
overlapMatrix(cms)

}
