% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-pathways.R
\name{pathways}
\alias{pathways}
\alias{pathways,ConsortiumMetabolism-method}
\alias{pathways,ConsortiumMetabolismSet-method}
\alias{pathways,ConsortiumMetabolismAlignment-method}
\title{Retrieve Metabolic Pathways}
\usage{
pathways(object, ...)

\S4method{pathways}{ConsortiumMetabolism}(object)

\S4method{pathways}{ConsortiumMetabolismSet}(
  object,
  type = c("all", "pan-cons", "niche", "core", "aux"),
  quantileCutoff = 0.1
)

\S4method{pathways}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolism} object.}

\item{...}{Object specific arguments.}

\item{type}{Character scalar giving the type of edges to output.}

\item{quantileCutoff}{Numeric scalar between 0 and 1 giving the
quantile threshold to use for filtering edges. For "pan-cons"
and "core" types, edges above \code{1 - quantileCutoff} are
returned. For "niche" and "aux" types, edges below
\code{quantileCutoff} are returned. Defaults to 0.1
(i.e., top/bottom 10 percent).}
}
\value{
A tibble containing pathway information
including:
\itemize{
\item consumed/produced metabolites
\item number of species involved
\item consumption/production sums
\item effective consumption/production metrics
}
}
\description{
Retrieves the pathways representing metabolic
interactions between species.
The argument \code{type} can be used to return only specific
types of pathways from a \code{ConsortiumMetabolismSet} type
object.
\itemize{
\item \code{all} will return all pathways
\item \code{pan-cons} will return only pathways that exist
in all consortia
\item \code{niche} will return niche pathways specific to
individual consortia
\item \code{core} will return core metabolic pathways
\item \code{aux} will return auxiliary pathways
}
}
\section{Methods (by class)}{
\itemize{
\item \code{pathways(ConsortiumMetabolism)}: Get Edges From a
\code{ConsortiumMetabolism} Object

\item \code{pathways(ConsortiumMetabolismSet)}: Get Edges From a
\code{ConsortiumMetabolismSet} Object

\item \code{pathways(ConsortiumMetabolismAlignment)}: Get Edges From a
\code{ConsortiumMetabolismAlignment} Object

}}
\examples{
cm <- synCM("test", n_species = 3, max_met = 5)
pathways(cm)

}
