% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plot,ConsortiumMetabolism,ANY-method}
\alias{plot,ConsortiumMetabolism,ANY-method}
\title{Plot a ConsortiumMetabolism object}
\usage{
\S4method{plot}{ConsortiumMetabolism,ANY}(
  x,
  type = c("Binary", "nSpecies", "Consumption", "Production", "EffectiveConsumption",
    "EffectiveProduction")
)
}
\arguments{
\item{x}{A \code{ConsortiumMetabolism} object.}

\item{type}{Character specifying the assay to plot.}
}
\value{
Invisibly returns \code{NULL}. Called for its
side effect of producing an igraph network plot.
}
\description{
Plot a ConsortiumMetabolism object
}
\examples{
cm <- synCM("test", n_species = 3, max_met = 5)
plot(cm)
}
