% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plot,ConsortiumMetabolismAlignment,ANY-method}
\alias{plot,ConsortiumMetabolismAlignment,ANY-method}
\title{Plot a ConsortiumMetabolismAlignment object}
\usage{
\S4method{plot}{ConsortiumMetabolismAlignment,ANY}(x, type = NULL)
}
\arguments{
\item{x}{A \code{ConsortiumMetabolismAlignment} object.}

\item{type}{Character specifying the plot type:
\code{"heatmap"}, \code{"network"}, or
\code{"scores"}.}
}
\value{
A \code{ggplot} object (returned invisibly).
}
\description{
Plot a ConsortiumMetabolismAlignment object
}
\examples{
\donttest{
cm1 <- synCM("comm_1", n_species = 3, max_met = 5)
cm2 <- synCM("comm_2", n_species = 4, max_met = 6)
cma <- align(cm1, cm2)
plot(cma)
}
}
