% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignmentHeatmap.R
\name{plotAlignmentHeatmap}
\alias{plotAlignmentHeatmap}
\title{MiCoAl Heatmap}
\usage{
plotAlignmentHeatmap(object, top = NULL, bottom = NULL)
}
\arguments{
\item{object}{An object of class MiCoAl.}

\item{top}{Character scalar giving the top fraction of consortia in which a
pathway should be present to be output in the heatmap. Defaults to NULL, in
which case all are plotted.}

\item{bottom}{Character scalar giving the bottom fraction of consortia in
which a pathway should be present to be output in the heatmap. Defaults to
NULL, in which case all are plotted.}
}
\value{
A ggplot heatmap
}
\description{
MiCoAl Heatmap
}
