% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramen-package.R
\docType{package}
\name{ramen-package}
\alias{ramen}
\alias{ramen-package}
\title{ramen: Reconstruction and Alignment of Microbial Exchange Networks}
\description{
Reconstructs metabolic exchange networks from microbial community data and aligns them using functional overlap scores to quantify the similarity and dissimilarity of consortium-level metabolic function. Provides tools for visualization and comparison of microbial consortia from a network perspective.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://admarhi.github.io/ramen/}
  \item Report bugs at \url{https://github.com/admarhi/ramen/issues}
}

}
\author{
\strong{Maintainer}: Adrian Hirt \email{am.hirt@pm.me} (\href{https://orcid.org/0009-0008-1929-6629}{ORCID})

}
\keyword{internal}
