% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramen-package.R
\docType{package}
\name{ramen-package}
\alias{ramen}
\alias{ramen-package}
\title{ramen: Reconstruction and Alignment of Microbial Exchange Networks}
\description{
Analysis of the functional (dis-) similarity of microbial communities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://admarhi.github.io/ramen/}
}

}
\author{
\strong{Maintainer}: Adrian Hirt \email{am.hirt@pm.me} (\href{https://orcid.org/0009-0008-1929-6629}{ORCID})

}
\keyword{internal}
