% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAMEN-package.R
\docType{package}
\name{RAMEN-package}
\alias{RAMEN}
\alias{RAMEN-package}
\title{RAMEN: RAMEN: Regional Association of Methylome variability with the Exposome and geNome}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R package that identifies which genetic (G), environmental (E), additive (G+E) or interaction (GxE) effect better explains DNA methylation levels in Variable Methylated Loci using microarray data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ericknavarrod.github.io/RAMEN/}
  \item Report bugs at \url{https://github.com/ErickNavarroD/RAMEN/issues}
}

}
\author{
\strong{Maintainer}: Erick I. Navarro-Delgado \email{erick.delgado@ubc.ca} (\href{https://orcid.org/0000-0003-1040-3519}{ORCID})

Authors:
\itemize{
  \item Keegan Korthauer \email{keegan@stat.ubc.ca} (\href{https://orcid.org/0000-0002-4565-1654}{ORCID})
  \item Michael S. Kobor \email{michael.kobor@ubc.ca} (\href{https://orcid.org/0000-0003-4140-1743}{ORCID})
}

}
\keyword{internal}
