% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAMEN-package.R
\docType{package}
\name{RAMEN-package}
\alias{RAMEN}
\alias{RAMEN-package}
\title{RAMEN: RAMEN: Regional Association of Methylome variability with the Exposome and geNome}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R package that identifies which genetic (G), environmental (E), additive (G+E) or interaction (GxE) effect better explains DNA methylation levels in Variable Methylated Loci using microarray data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ericknavarrod.github.io/RAMEN/}
}

}
\author{
\strong{Maintainer}: Erick I. Navarro-Delgado \email{ericknadel98@hotmail.com} (\href{https://orcid.org/0000-0003-1040-3519}{ORCID})

}
\keyword{internal}
