% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAMEN-package.R
\docType{package}
\name{RAMEN-package}
\alias{RAMEN}
\alias{RAMEN-package}
\title{RAMEN: Regional Association of Methylome variability with the Exposome and geNome}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

RAMEN is an R package which goal is to identify Variable Methylated Loci (VML) in microarray DNA methylation data and then, using genome and exposome data, identify which individual genetic (G) or environmental (E) or joint (G+E or GxE) model better explains the VML variability. This package provides estimates of genetic and environmental contribution for each VML, allowing researchers to gain insights into the complex interplay between genetics, environment and DNA methylation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropensci.github.io/RAMEN/} 
  \item \url{https://github.com/ropensci/RAMEN/}
  \item Report bugs at \url{https://github.com/ropensci/RAMEN/issues}
}

}
\author{
\strong{Maintainer}: Erick I. Navarro-Delgado \email{erick.delgado@ubc.ca} (\href{https://orcid.org/0000-0003-1040-3519}{ORCID})

Authors:
\itemize{
  \item Erick I. Navarro-Delgado \email{erick.delgado@ubc.ca} (\href{https://orcid.org/0000-0003-1040-3519}{ORCID})
  \item Keegan Korthauer \email{keegan@stat.ubc.ca} (\href{https://orcid.org/0000-0002-4565-1654}{ORCID})
  \item Michael S. Kobor \email{michael.kobor@ubc.ca} (\href{https://orcid.org/0000-0003-4140-1743}{ORCID})
}

Other contributors:
\itemize{
  \item Lluís Revilla Sancho \email{lluis.revilla@gmail.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID}) [reviewer]
  \item Ulduz Vafadarshamasbi \email{vafadar.u@uq.edu.au} (\href{https://orcid.org/0000-0002-8370-8084}{ORCID}) [reviewer]
  \item The University of British Columbia [funder]
}

}
\keyword{internal}
