% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-setName.R
\name{setName}
\alias{setName}
\alias{setName,ConsortiumMetabolism-method}
\alias{setName,ConsortiumMetabolismSet-method}
\alias{setName,ConsortiumMetabolismAlignment-method}
\title{Set Names or description of ramen Objects}
\usage{
setName(object, value = NA_character_)

\S4method{setName}{ConsortiumMetabolism}(object, value = NA_character_)

\S4method{setName}{ConsortiumMetabolismSet}(object, value = NA_character_)

\S4method{setName}{ConsortiumMetabolismAlignment}(object, value = NA_character_)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolism},
\code{ConsortiumMetabolismSet}, or \code{ConsortiumMetabolismAlignment}
object}

\item{value}{Character scalar specifying a name.}
}
\value{
A cluster from a \code{ConsortiumMetabolismSet} object
}
\description{
Sets a new name or description of \code{ConsortiumMetabolism},
\code{ConsortiumMetabolismSet}, or \code{ConsortiumMetabolismAlignment}
objects
}
