% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-show.R
\name{show,ConsortiumMetabolismAlignment-method}
\alias{show,ConsortiumMetabolismAlignment-method}
\title{Show method for \code{ConsortiumMetabolismAlignment} Objects}
\usage{
\S4method{show}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{a \code{ConsortiumMetabolismAlignment} object.}
}
\description{
Show method for \code{ConsortiumMetabolismAlignment} Objects
}
