% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cma.R
\name{similarityMatrix}
\alias{similarityMatrix}
\alias{similarityMatrix,ConsortiumMetabolismAlignment-method}
\title{Get Similarity Matrix}
\usage{
similarityMatrix(object)

\S4method{similarityMatrix}{ConsortiumMetabolismAlignment}(object)
}
\arguments{
\item{object}{A \link{ConsortiumMetabolismAlignment} object
of type \code{"multiple"}.}
}
\value{
A numeric n x n matrix.
}
\description{
Returns the pairwise similarity matrix from a
multiple alignment.
}
\section{Methods (by class)}{
\itemize{
\item \code{similarityMatrix(ConsortiumMetabolismAlignment)}: Similarity matrix from a
\link{ConsortiumMetabolismAlignment}. For multiple alignments
this is the symmetric n x n pairwise matrix; for database
search alignments it is a 1 x n row vector of the query's
scores against each database member.

}}
\examples{
cm1 <- synCM("comm_1", n_species = 3, max_met = 5)
cm2 <- synCM("comm_2", n_species = 4, max_met = 6)
cms <- ConsortiumMetabolismSet(cm1, cm2, name = "test")
cma <- align(cms)
similarityMatrix(cma)

}
