% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-accessors-cm.R,
%   R/methods-accessors-cma.R, R/methods-accessors-cms.R
\name{species}
\alias{species}
\alias{species,ConsortiumMetabolism-method}
\alias{species,ConsortiumMetabolismAlignment-method}
\alias{species,ConsortiumMetabolismSet-method}
\title{Return Species in a Consortium}
\usage{
species(object, ...)

\S4method{species}{ConsortiumMetabolism}(object)

\S4method{species}{ConsortiumMetabolismAlignment}(object)

\S4method{species}{ConsortiumMetabolismSet}(object, ...)
}
\arguments{
\item{object}{A \code{ConsortiumMetabolismSet} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A character vector of species names.
}
\description{
Returns the species present in a consortium or set of
consortia.
}
\section{Methods (by class)}{
\itemize{
\item \code{species(ConsortiumMetabolism)}: Return Species in a
\code{ConsortiumMetabolism}

\item \code{species(ConsortiumMetabolismAlignment)}: Return species from a
\code{ConsortiumMetabolismAlignment}

\item \code{species(ConsortiumMetabolismSet)}: Return Species in a
\code{ConsortiumMetabolismSet}

}}
\examples{
cm <- synCM("test", n_species = 3, max_met = 5)
species(cm)

}
