% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_methylation_data.R
\docType{data}
\name{test_methylation_data}
\alias{test_methylation_data}
\title{Methylation data matrix example}
\format{
\subsection{\code{test_methylation_data}}{

A data frame with 3,000 rows and 30 columns:
\describe{
\item{\emph{rownames}}{Probe IDs (column TargetID in the EPIC array)}
\item{ID1:30}{DNAme profile of individuals 1 to 30; column names correspond to individual IDs}
...
}
}
}
\usage{
test_methylation_data
}
\description{
Simulated M values of the 3000 probes selected in \emph{test_array_manifest} for 30 individuals. Values were converted from beta values, which were drawn from a bimodal Beta distribution.
}
\keyword{datasets}
