% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultrastable_cpgs.R
\docType{data}
\name{ultrastable_cpgs}
\alias{ultrastable_cpgs}
\title{Ultrastable probes}
\format{
\subsection{\code{ultrastable_cpgs}}{

A vector with XXXXX
\describe{
\item{\emph{rownames}}{Probe ID - for storage reasons, this variable was stored as row names, but rownames have to be converted to a new column called "TargetID" prior to its use in RAMEN.}
\item{MAPINFO}{Probe genomic position (h19)}
\item{CHR}{Chromosome}
\item{STRAND}{Strand}
...
}
}
}
\source{
XX
}
\usage{
ultrastable_cpgs
}
\description{
A DESCRIPTION OF THE DATASET
}
\keyword{datasets}
